Fix note delete command NumberFormatException

+++ src/main/java/seedu/address/logic/commands/note/NoteDeleteCommand.java
@@ -15,6 +15,7 @@ import seedu.address.seplendidui.UiUtil;
  * Deletes a note to the NoteList.
  */
 public class NoteDeleteCommand extends NoteCommand {
+    public static final String MESSAGE_NONEXISTENT_NOTE = "Note not found, please put a valid index.";
     public static final String ACTION_WORD = "delete";
     public static final String MESSAGE_SUCCESS = "Note deleted: %1$s";
 
@@ -46,6 +47,12 @@ public class NoteDeleteCommand extends NoteCommand {
     @Override
     public CommandResult execute(SeplendidModel seplendidModel) throws CommandException {
         requireNonNull(seplendidModel);
+
+        int noteListSize = seplendidModel.getNoteCatalogue().getNoteList().size();
+
+        if (this.noteIndexToDelete == -1 || noteListSize < this.noteIndexToDelete) {
+            throw new CommandException(MESSAGE_NONEXISTENT_NOTE);
+        }
 
         Note deletedNote = seplendidModel.deleteNote(noteIndexToDelete);
         return new CommandResult(String.format(MESSAGE_SUCCESS, Messages.format(deletedNote)),
